rename: 컴포넌트 안에 페이지별 폴더 생성 후 파일 이동
