+++ src/components/main/ProductItem.tsx
@@ -1,9 +1,9 @@
 import { ChangeEventHandler } from 'react';
 import { css, styled } from 'styled-components';
-import { NOT_NUMBER, QUANTITY, MAX_NUMBER_LENGTH } from '../constants';
-import { useSetCart } from '../hooks/useCart';
-import { useLoadCart } from '../hooks/useLoadCart';
-import CartIcon from './icons/CartIcon';
+import { NOT_NUMBER, QUANTITY, MAX_NUMBER_LENGTH } from '../../constants';
+import { useSetCart } from '../../hooks/useCart';
+import { useLoadCart } from '../../hooks/useLoadCart';
+import CartIcon from '../icons/CartIcon';
 import QuantityInput from './QuantityInput';
 
 interface Props {
