+++ src/pages/MainPage.tsx
@@ -1,5 +1,5 @@
 import Header from '../components/common/Header';
-import ProductList from '../components/ProductList';
+import ProductList from '../components/main/ProductList';
 
 const MainPage = () => {
   return (
